# gearbox-agent/docs/swagger.yaml
basePath: /
definitions:
  api.GitHubPushPayload:
    properties:
      head_commit:
        properties:
          id:
            type: string
          message:
            type: string
        type: object
      pusher:
        properties:
          name:
            type: string
        type: object
      ref:
        type: string
      repository:
        properties:
          full_name:
            type: string
        type: object
    type: object
  api.HealthResponse:
    properties:
      status:
        example: ok
        type: string
      timestamp:
        example: "2024-01-17T10:30:00Z"
        type: string
      uptime:
        example: 2h30m15s
        type: string
      version:
        example: 1.0.0
        type: string
    type: object
  api.MetadataResponse:
    properties:
      last_error:
        type: string
      last_sync_time:
        type: string
      metadata:
        $ref: '#/definitions/haproxy.Metadata'
    type: object
  api.SyncStatusResponse:
    properties:
      backend_count:
        type: integer
      last_error:
        type: string
      last_sync_time:
        type: string
    type: object
  api.WSInfoResponse:
    properties:
      enabled:
        example: true
        type: boolean
      endpoint:
        example: /api/v1/events
        type: string
      event_types:
        example:
        - sync.started
        - sync.completed
        - sync.failed
        - config.changed
        - webhook.received
        items:
          type: string
        type: array
      subscribers:
        example: 2
        type: integer
    type: object
  api.WSTokenResponse:
    properties:
      expires_in:
        description: Seconds until expiry
        example: 60
        type: integer
      token:
        example: a1b2c3d4e5f6789012345678901234567890abcdef1234567890abcdef123456
        type: string
    type: object
  api.WebhookErrorResponse:
    properties:
      error:
        example: invalid_signature
        type: string
      message:
        example: Signature verification failed
        type: string
      status:
        example: error
        type: string
    type: object
  api.WebhookInfoResponse:
    properties:
      enabled:
        example: true
        type: boolean
      message:
        example: Configure this URL in GitHub repository settings
        type: string
      webhook_url:
        example: https://proxy.example.com:8405/api/v1/webhook/github
        type: string
    type: object
  api.WebhookSuccessResponse:
    properties:
      message:
        example: Sync triggered
        type: string
      status:
        example: accepted
        type: string
    type: object
  haproxy.ACLMetadata:
    properties:
      type:
        type: string
      value:
        type: string
    type: object
  haproxy.ActiveSession:
    properties:
      age_seconds:
        description: How long the session has been active
        type: integer
      backend:
        type: string
      bytes_in:
        type: integer
      bytes_out:
        type: integer
      frontend:
        type: string
      id:
        type: string
      idle_seconds:
        description: How long since last activity
        type: integer
      request_count:
        description: Number of requests in this session
        type: integer
      server:
        type: string
      source_ip:
        type: string
      source_port:
        type: integer
    type: object
  haproxy.BackendMetadata:
    properties:
      acls:
        items:
          $ref: '#/definitions/haproxy.ACLMetadata'
        type: array
      app_name:
        type: string
      containers:
        items:
          $ref: '#/definitions/haproxy.ContainerMetadata'
        type: array
      frontend:
        type: string
      hostname:
        type: string
      name:
        type: string
      public:
        type: boolean
      rate_limit:
        type: integer
      servers:
        items:
          type: string
        type: array
      ssl_enabled:
        type: boolean
      ssl_verify:
        type: string
    type: object
  haproxy.ConnectionFlow:
    properties:
      active_connections:
        type: integer
      backend:
        type: string
      bytes_in:
        type: integer
      bytes_out:
        type: integer
      source_ip:
        type: string
      total_requests:
        type: integer
    type: object
  haproxy.ContainerMetadata:
    properties:
      depends_on:
        items:
          type: string
        type: array
      is_backend:
        type: boolean
      name:
        type: string
      network_mode:
        type: string
    type: object
  haproxy.FrontendMetadata:
    properties:
      backends:
        items:
          type: string
        type: array
      name:
        type: string
    type: object
  haproxy.Metadata:
    properties:
      backends:
        items:
          $ref: '#/definitions/haproxy.BackendMetadata'
        type: array
      frontends:
        items:
          $ref: '#/definitions/haproxy.FrontendMetadata'
        type: array
      generated_at:
        type: string
      version:
        type: string
    type: object
  traffic.BackendTrafficData:
    properties:
      avg_response_time:
        type: integer
      bytes_in:
        type: integer
      bytes_out:
        type: integer
      error_rate:
        type: number
      name:
        type: string
      response_2xx:
        type: integer
      response_3xx:
        type: integer
      response_4xx:
        type: integer
      response_5xx:
        type: integer
      total_requests:
        type: integer
      unique_ips:
        type: integer
    type: object
  traffic.TrafficResponse:
    properties:
      active_sessions:
        items:
          $ref: '#/definitions/haproxy.ActiveSession'
        type: array
      backend_traffic:
        items:
          $ref: '#/definitions/traffic.BackendTrafficData'
        type: array
      collected_at:
        type: string
      connection_flows:
        items:
          $ref: '#/definitions/haproxy.ConnectionFlow'
        type: array
      sources:
        items:
          $ref: '#/definitions/traffic.TrafficSourceData'
        type: array
      top_by_bytes:
        items:
          $ref: '#/definitions/traffic.TrafficSourceData'
        type: array
      top_by_connections:
        items:
          $ref: '#/definitions/traffic.TrafficSourceData'
        type: array
      top_by_requests:
        items:
          $ref: '#/definitions/traffic.TrafficSourceData'
        type: array
      total_sources:
        type: integer
    type: object
  traffic.TrafficSourceData:
    properties:
      backend:
        type: string
      bytes_in:
        type: integer
      bytes_out:
        type: integer
      connections:
        type: integer
      frontend:
        type: string
      http_request_rate:
        description: HTTP requests per period
        type: integer
      ip_address:
        type: string
      last_seen:
        type: string
      requests:
        type: integer
      session_rate:
        description: Sessions per period
        type: integer
    type: object
  traffic.TrafficSummary:
    properties:
      avg_response_time:
        type: integer
      backend_count:
        type: integer
      collected_at:
        type: string
      error_rate:
        type: number
      frontend_count:
        type: integer
      total_bytes_in:
        type: integer
      total_bytes_out:
        type: integer
      total_requests:
        type: integer
      total_response_2xx:
        type: integer
      total_response_3xx:
        type: integer
      total_response_4xx:
        type: integer
      total_response_5xx:
        type: integer
      unique_ips:
        type: integer
    type: object
host: localhost:8405
info:
  contact:
    name: Gearbox Agent Support
    url: https://github.com/sarg3nt/gearbox-agent/issues
  description: REST API for monitoring and managing HAProxy instances. Provides endpoints
    for system metrics, HAProxy stats, log streaming, security monitoring, and real-time
    WebSocket events.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/sarg3nt/gearbox-agent
  title: Gearbox Agent API
  version: "1.0"
paths:
  /api/v1/events:
    get:
      description: 'Upgrades to WebSocket connection for real-time event streaming.
        First obtain a token via POST /api/v1/events/token, then connect with ?token=<token>.
        Events: sync.started, sync.completed, sync.failed, config.changed, webhook.received.'
      parameters:
      - description: Short-lived WebSocket token from /api/v1/events/token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols - WebSocket connection established
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: WebSocket events stream
      tags:
      - WebSocket
  /api/v1/events/info:
    get:
      description: Returns information about the WebSocket endpoint including available
        event types and current subscriber count.
      produces:
      - application/json
      responses:
        "200":
          description: WebSocket endpoint info
          schema:
            $ref: '#/definitions/api.WSInfoResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: WebSocket info
      tags:
      - WebSocket
  /api/v1/events/token:
    post:
      description: Exchanges an API key for a short-lived (60 second) single-use token
        for WebSocket authentication. Use the returned token in the WebSocket connection
        query parameter.
      produces:
      - application/json
      responses:
        "200":
          description: WebSocket token
          schema:
            $ref: '#/definitions/api.WSTokenResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Failed to generate token
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Exchange API key for WebSocket token
      tags:
      - WebSocket
  /api/v1/metadata:
    get:
      description: Returns HAProxy configuration metadata including frontends, backends,
        and their settings from the last sync.
      produces:
      - application/json
      responses:
        "200":
          description: HAProxy metadata
          schema:
            $ref: '#/definitions/api.MetadataResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "503":
          description: Sync service not configured or pending
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: HAProxy metadata
      tags:
      - Sync
  /api/v1/sync/status:
    get:
      description: Returns the status of the Git sync service including last sync
        time, error status, and backend count.
      produces:
      - application/json
      responses:
        "200":
          description: Sync status
          schema:
            $ref: '#/definitions/api.SyncStatusResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "503":
          description: Sync service not configured
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Sync status
      tags:
      - Sync
  /api/v1/traffic:
    get:
      description: Returns traffic analysis data including per-IP statistics from
        stick tables and backend aggregations.
      parameters:
      - default: 100
        description: Limit number of sources returned
        in: query
        name: limit
        type: integer
      - default: 10
        description: Number of top sources to return
        in: query
        name: top_n
        type: integer
      - default: false
        description: Include all IPs, not just those with data
        in: query
        name: include_all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Traffic analysis data
          schema:
            $ref: '#/definitions/traffic.TrafficResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to fetch traffic data
          schema:
            type: string
        "503":
          description: HAProxy stats not configured
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Traffic analysis
      tags:
      - Traffic
  /api/v1/traffic/summary:
    get:
      description: Returns a summary of traffic statistics without individual IP details.
      produces:
      - application/json
      responses:
        "200":
          description: Traffic summary
          schema:
            $ref: '#/definitions/traffic.TrafficSummary'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Failed to fetch traffic data
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Traffic summary
      tags:
      - Traffic
  /api/v1/webhook/github:
    post:
      consumes:
      - application/json
      description: Receives GitHub push events and triggers a sync. Authenticated
        via HMAC-SHA256 signature in X-Hub-Signature-256 header.
      parameters:
      - description: GitHub HMAC signature
        in: header
        name: X-Hub-Signature-256
        required: true
        type: string
      - description: GitHub event type
        in: header
        name: X-GitHub-Event
        required: true
        type: string
      - description: GitHub push payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.GitHubPushPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Webhook processed successfully
          schema:
            $ref: '#/definitions/api.WebhookSuccessResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/api.WebhookErrorResponse'
        "401":
          description: Invalid or missing signature
          schema:
            $ref: '#/definitions/api.WebhookErrorResponse'
      summary: GitHub webhook
      tags:
      - Webhook
  /api/v1/webhook/info:
    get:
      description: Returns webhook configuration information including the URL to
        configure in GitHub.
      produces:
      - application/json
      responses:
        "200":
          description: Webhook configuration info
          schema:
            $ref: '#/definitions/api.WebhookInfoResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Webhook info
      tags:
      - Webhook
  /health:
    get:
      description: Returns the health status of the gearbox-agent service. No authentication
        required.
      produces:
      - application/json
      responses:
        "200":
          description: Service is healthy
          schema:
            $ref: '#/definitions/api.HealthResponse'
      summary: Health check
      tags:
      - Health
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'API key authentication. Use format: Bearer YOUR_API_KEY'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"